Split OpenAI loop backend tests by responsibility
Preserve all 83 examples and their execution order while extracting codec, state, recovery, and transport specs plus shared fixtures.
